Move get_next_available_network_number() function to utils
